MaquinadeConversaoV2/requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
streamlit==1.35.0
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1

# ── AI / LLM ──────────────────────────────────────────────────────────────────
groq==0.9.0

# ── TTS ───────────────────────────────────────────────────────────────────────
elevenlabs==1.2.2
edge-tts>=6.1.9

# ── Media Processing ──────────────────────────────────────────────────────────
moviepy==1.0.3
ffmpeg-python==0.2.0
Pillow==10.3.0
numpy==1.26.4
imageio==2.34.1
imageio-ffmpeg==0.5.1

# ── Whisper (local) ───────────────────────────────────────────────────────────
openai-whisper==20231117
torch==2.3.0

# ── Audio ─────────────────────────────────────────────────────────────────────
pydub==0.25.1

# ── Google APIs ───────────────────────────────────────────────────────────────
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# ── HTTP / Networking ─────────────────────────────────────────────────────────
httpx==0.27.0
requests==2.32.2
pyngrok==7.2.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-multipart==0.0.9
aiofiles==23.2.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-mock==3.14.0
